Misc. - String Floats for Numeric Column Value (#206)
Currently if the value of a `NumberValue` is a string representation of a float (e.g. "95.0") then the `_isint()` function returns `True` and the `_convert()` function tries to cast it to an int. However (in Python 3.8 at least) you can't cast "95.0" directly to an `int` - you have to cast it to a `float` first.